refactor: split pseudo_etfs.py router into CRUD and analysis modules (#95)
