+++ backend/app/main.py
@@ -13,7 +13,7 @@ from sqlalchemy import select, text
 
 from app.database import async_session, engine, Base
 from app.models import Asset  # noqa: F401 - ensure models are imported for create_all
-from app.routers import annotations, assets, groups, holdings, portfolio, prices, pseudo_etfs, quotes, settings as settings_router, tags, thesis, watchlist
+from app.routers import annotations, assets, groups, holdings, portfolio, prices, pseudo_etfs, pseudo_etf_analysis, quotes, settings as settings_router, tags, thesis, watchlist
 from app.services.price_sync import sync_all_prices
 from app.services.yahoo import batch_fetch_currencies
 
@@ -180,6 +180,7 @@ app.include_router(holdings.router)
 app.include_router(thesis.router)
 app.include_router(annotations.router)
 app.include_router(pseudo_etfs.router)
+app.include_router(pseudo_etf_analysis.router)
 app.include_router(quotes.router)
 app.include_router(settings_router.router)
 app.include_router(watchlist.router)
